trim: Don't trim user-inserted debugging calls as having no side effect
While a call like glPushDebugGroup or glStringMarkerGREMEDY may
technically not effect the OpenGL state, users put these calls in
place specifically to be able to get more information out of tools
such as apitrace. So we definitely don't want to trim these calls out
when a user requests trimming or pruning calls with no side effects.